+++ CHANGELOG.md
@@ -5,4 +5,5 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 - unit tests
+- randomly encryption error using openssl (see: https://stackoverflow.com/questions/37439981/openssl-encrypt-randomly-fails-iv-passed-is-only-x-bytes-long-cipher-exp?answertab=votes#tab-top)
 - removed container argument from `RemoteImageCache` constructor